DPSPS-126: Rework resend with a different name bug
The whole solution here is a little poor but we effectively want to resend emails
with the previous content but new names and we don't have access to the original
template or original arguments. This "slightly better than the last one" implementation
allows the html to carry name placeholders that we can substitute rather than relying
on the text "Dear x". The HtmlStripper class has been moved here from rtp-caseworker-domain-lib
so it's available to the resend command.