DPSPS-126: Very poor refix of name updates when resending emails
Ideally we'd regenerate the email again but other problems and time
constraints mean this hasn't really panned out. We've decided this
is acceptable, although not ideal, in the very short term. The problem
with this logic is it depends on knowledge of email templates it
shouldn't know and that it uses regexs in a way that allows manipulation
of the HTML as well as just breaking the content itself.